refactor: move rename 'updates' category to 'admin'
