+++ docs/pages/commands.mdx
@@ -37,6 +37,6 @@ These server-specific commands rely on server-level data and interactions, which
 <Cards>
   <Card title="Guides" icon={<FaMapMarkedAlt />} href="/commands/guides" />
   <Card title="Informations" icon={<HiMiniInformationCircle />} href="/commands/info" />
-  <Card title="Updates" icon={<IoIosNotifications />} href="/commands/updates" />
+  <Card title="Admin" icon={<IoIosNotifications />} href="/commands/admin" />
   <Card title="Utilities" icon={<IoCog />} href="/commands/utility" />
 </Cards>
