+++ docs/pages/commands/_meta.json
@@ -1,6 +1,6 @@
 {
   "guides": "Guides",
   "info": "Informations",
-  "updates": "Updates",
+  "admin": "Admin",
   "utility": "Utilities"
 }
