+++ docs/pages/commands/admin.mdx
@@ -1,19 +1,19 @@
 ---
-title: Updates Commands
-description: A comprehensive guide on live updates commands offered by SkyHelper bot
+title: Admin Commands
+description: All the admin commands offered by SkyHelper
 ---
 
 import { Callout, Steps } from "nextra/components";
 import { ImFire } from "react-icons/im";
 import { CiWarning } from "react-icons/ci";
 
-# Updates Commands
+# Admin Commands
 
 Almost all commands under this category are related to features providing live updates and/or reminders.
 
 <Callout type="error" emoji={<ImFire />}>
   All commands in this category are only available for servers (not [user-apps](/pages/user-install-app)) and require users to
-  have `Manage Server` permissions to use them. These commands are also not visible to users who lack the permission (this
+  have `Manage Server` permissions(Admins) to use them. These commands are also not visible to users who lack the permission (this
   behavior can be adjusted under `Server Settings > Integrations Tab`, although there’s also an internal check for permissions, so
   enabling it for users without the necessary permission will still prevent them from using it).
 </Callout>
